+++ python_secrets/utils.py
@@ -8,16 +8,18 @@ import textwrap
 
 from cliff.command import Command
 from cliff.lister import Lister
+from configobj import ConfigObj
 from os import listdir, sep
 from os.path import abspath, basename, isdir
 from six.moves import input
 
 
+AWS_CONFIG_FILE = os.path.join(os.path.expanduser('~'), '.aws', 'credentials')
 OPENDNS_URL = 'https://diagnostic.opendns.com/myip'
+
 # NOTE: While calling subprocess.call() with shell=True can have security
 # implications, the person running this command already has control of her
 # account.
-
 
 def get_output(cmd=['echo', 'NO COMMAND SPECIFIED'],
                stderr=subprocess.STDOUT,
@@ -246,5 +248,46 @@ def tree(dir, padding='', print_files=True, isLast=False, isFirst=True):
             else:
                 print(padding + '├── ' + file)
 
+
+class SetAWSCredentials(Command):
+    """Set credentials from saved secrets for use by AWS CLI.
+
+    This command directly manipulates the AWS CLI "credentials" INI-style file.
+    The AWS CLI does not support non-interactive manipulation of the credentials
+    file, so this hack is used to do this. Be aware that this might cause some
+    problems (though it shouldn't, since the file is so simple.)
+
+    Use the --user option to select a specific user, otherwise "default" is
+    used.
+    """
+
+    # See https://docs.aws.amazon.com/cli/latest/userguide/cli-config-files.html
+
+    log = logging.getLogger(__name__)
+
+    def get_parser(self, prog_name):
+        parser = super().get_parser(prog_name)
+        parser.add_argument(
+            '-U', '--user',
+            action='store',
+            dest='user',
+            default='default',
+            help='IAM User who owns credentials (default: "default")'
+        )
+        return parser
+
+    def take_action(self, parsed_args):
+        self.log.debug('setting AWS CLI IAM user credentials')
+        self.app.secrets.requires_environment()
+        self.app.secrets.read_secrets_and_descriptions()
+        required_vars = ['aws_access_key_id', 'aws_secret_access_key']
+        config = ConfigObj(AWS_CONFIG_FILE)
+        for v in required_vars:
+            try:
+                cred = self.app.secrets.get_secret(v)
+            except Exception as e:
+                raise
+            config[parsed_args.user][v] = cred
+        config.write()
 
 # vim: set fileencoding=utf-8 ts=4 sw=4 tw=0 et :